requirements.txt
# 医療 AI 文献トレンド分析アプリ 依存関係リスト

# ML/深層学習
torch==2.9.0
numpy==1.26.4

# データ処理/時系列
pandas==2.2.2
python-dateutil==2.9.0

# API / バックエンド
fastapi==0.111.0
uvicorn==0.30.1
flask==3.0.3
requests==2.31.0
pydantic==2.7.3

# フロントエンド / 可視化
streamlit==1.35.0
plotly==5.22.0

# データサイエンス / トレンド分析
scikit-learn==1.3.2
joblib==1.4.2
nltk==3.8.1

# 環境管理
python-dotenv==1.0.1